+++ src/factorio/factories.py
@@ -2,8 +2,10 @@ from __future__ import annotations
 
 from inspect import getmro
 from typing import Any, Generic, TypeVar
+from warnings import warn
 
 from factorio.fields import AbstractField
+from factorio.lib.exceptions import RemovedIn07Warning
 
 T = TypeVar("T")
 
@@ -31,6 +33,25 @@ class Factory(Generic[T]):
 
     @classmethod
     def build(cls, **kwargs: Any) -> T:
+        if hasattr(cls, "Fields"):
+            return cls._old_style_build(**kwargs)
+
+        model = cls.get_model()
+        fields: dict[str, Any] = {}
+        for key, value in cls.__dict__.items():
+            if isinstance(value, AbstractField):
+                fields[key] = value()
+        for key, value in kwargs.items():
+            fields[key] = value() if isinstance(value, AbstractField) else value
+
+        return model(**fields)
+
+    @classmethod
+    def _old_style_build(cls, **kwargs: Any) -> T:
+        warning = RemovedIn07Warning(
+            "Factories with `Fields` class", "ConstantField for constant values"
+        )
+        warn(warning, stacklevel=3)
         model = cls.get_model()
         fields: dict[str, Any] = kwargs
         for key, value in cls.Fields.__dict__.items():